\name{convertAllToNumeric}
\alias{convertAllToNumeric}
\alias{convertAllToNumeric-methods}
\alias{convertAllToNumeric,data.frame-method}
\alias{convertAllToNumeric,matrix-method}
\alias{convertAllToNumeric,numeric-method}
\alias{convertAllToNumeric,vector-method}
\title{
convert all to numeric values
}
\description{
convert all not numeric values to numeric or NA
thereby character convert to NAs also empty values like '' will be convert to NAs
}
\usage{
convertAllToNumeric(dataSet, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataSet}{
any data set
}
  \item{\dots}{
not used
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}

\seealso{
\code{\link{incReg}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
