\name{exampleIncRegData}
\alias{exampleIncRegData}
\docType{data}
\title{
experimental pkb values
}

\description{
example data set for the package incReg.
contents 36 substances with experimental pkb values and 6 descriptors calculated with chemoinformatics methods (MSINDO, PM6, GAUSSIAN)
}

\usage{
data(exampleIncRegData)
}

\format{
A data set with 36 substances and 8 columns
	\describe{
	\item{\code{Name}}{a character vector}
	\item{\code{exp pkb}}{a numeric vector}
	\item{\code{dipole D}}{a numeric vector}
	\item{\code{total energy eV}}{a numeric vector}
	\item{\code{Klopman one divided by loca au}}{a numeric vector}
	\item{\code{Loca energy eV}}{a numeric vector}
	\item{\code{charge of N e}}{a numeric vector}
	\item{\code{proton affinity kcal/mol}}{a numeric vector}
	
	}
}

\references{
Sanderson P (2003) Azaindoles: moderately basic P1 groups for enhancing the selectivity of thrombin inhibitors. Bioorg Med Chem Lett 13:795-798. doi: 10.1016/S0960-894X(03)00017-9
Perrin DD (1965) Prediction of the strengths of some organic bases. J Chem Soc 18:5590-5596. doi: 10.1039/jr9650005590
Albert A, Goldacre R, Phillips J (1948) 455. The Strength of Heterocyclic Bases. J Chem Soc 2240-2249. doi: 10.1039/JR9480002240
Catalan J, Elguero J (1983) Basicity of azoles. Part 2. Theoretical study of the basicity of methylpyrazoles and methylimidazoles. J Chem Soc Perkin Trans II 1869. doi: 10.1039/p29830001869
Catalan J, Menendez M, Laynez J, Claramunt RM, Bruix M, Mendoza J De, Elguero J (1985) Basicity of azoles. VII. Basicity of C-aminopyrazoles in relation to tautomeric and protonation studies. J Heterocyclic Chem 22:997-1000. doi: 10.1002/jhet.5570220413
Kostka K, Strawiak MM (1984) Protolytic and complexing properties of pyrazoles. Part III. 1-Phenylpyrazole, 3(5)-phenylpyrazole, 3,5-diphenylpyrazole, 1-methyl-3,5-dimethylpyrazole, and 1,3,5-triphenylpyrazole. Pol J Chem 58:647
Llamas-Saiz AL, Foces-Foces C, Cano FH, Jimenez P, Laynez J, Meutermans W, Elguero J, Limbach H-H, Aguilar-Parrilla F (1994) The influence of the nitro group on the solid-state structure of 4-nitropyrazoles: the cases of pyrazole, 3,5-dimethylpyrazole, 3,5-di- tert -butylpyrazole and 3,5-diphenylpyrazole. I. Static aspects (crystallography and thermodynamics). Acta Crystallogr B 50:746-762. doi: 10.1107/S0108768194004180
Jiang F, McCracken J, Peisach J (1990) Nuclear quadrupole interactions in copper(II)-diethylenetriamine-substituted imidazole complexes and in copper(II) proteins. J Am Chem Soc 112:9035-9044. doi: 10.1021/ja00181a002
Takeuchi Y, Kirk KL, Cohen LA (1978) Adjacent lone pair (ALP) effects in heteroaromatic systems. 2. Isotope exchange of ring hydrogens in nitro- and fluoroimidazoles. J Org Chem 43:3570-3578. doi: 10.1021/jo00412a034
ACD/Labs (2009) ACD/pKa. http://www.acdlabs.com/products/phys_chem_lab/pka/exp_table.html. Accessed 02 May 2009
Cassidy CS, Reinhardt LA, Cleland WW, Frey PA (1999) Hydrogen bonding in complexes of carboxylic acids with 1-alkylimidazoles: steric and isotopic effects on low barrier hydrogen bonding. J Chem Soc Perkin Trans II 635-642. doi: 10.1039/a806387g 
}

\keyword{datasets}
