\name{getAlphaValue}
\alias{getAlphaValue}
\alias{getAlphaValue-methods}
\alias{getCollResult}
\alias{getCollResult-methods}
\alias{getCsvFile}
\alias{getCsvFile-methods}
\alias{getCutlineAdvZeroTest}
\alias{getCutlineAdvZeroTest-methods}
\alias{getCutlineCorr}
\alias{getCutlineCorr-methods}
\alias{getCutlineMulticoll}
\alias{getCutlineMulticoll-methods}
\alias{getCutlineZeroTest}
\alias{getCutlineZeroTest-methods}
\alias{getDataSet}
\alias{getDataSet-methods}
\alias{getMissingValueIsCol}
\alias{getMissingValueIsCol-methods}
\alias{getMissingValueTyp}
\alias{getMissingValueTyp-methods}
\alias{getShouldCentering}
\alias{getShouldCentering-methods}
\alias{getShouldScaling}
\alias{getShouldScaling-methods}
\alias{getYname}
\alias{getYname-methods}
\alias{getMultiCollResult}
\alias{getMultiCollResult-methods}
\alias{getModelResult}
\alias{getModelResult-methods}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Getter functions
}
\description{
Returns the corresponding value from an incReg-Object
}
\usage{
getAlphaValue(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
represents an object of the class incReg
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{setAlphaValue<-}}
\code{\link{incReg}}

}

\keyword{manip}

