\name{getModelNumberX}
\alias{getModelNumberX}
\alias{getModelNumberX-methods}
\alias{getFinalModel}
\alias{getFinalModel-methods}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get result models
}
\description{
Return the result models of the incremental Multivariate Regression
}
\usage{
getModelNumberX(object, number)

getFinalModel(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
represents an object of the class incReg
}
  \item{number}{
number of the model which is produced during the incremental Multivariate Regression (last number is normally the final model)
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }

