\name{getPreModelResults}
\alias{getPreModelResults}
\alias{getPreModelResults-methods}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Multicollinearity and correlation results
}
\description{
Return the results of the multicollinearity and the correlation test
}
\usage{
getPreModelResults(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
represents an object of the class incReg
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}
}


\keyword{ misc }

