\name{IncRegFinalmodelResult-class}
\Rdversion{1.1}
\docType{class}
\alias{IncRegFinalmodelResult-class}

\title{Class \code{"IncRegFinalmodelResult"}}
\description{
Save the results of the incReg class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("IncRegFinalmodelResult", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{.multiCollResult}:}{Object of class \code{"list"} ~~ }
    \item{\code{.collResult}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{.modelResult}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "IncRegFinalmodelResult" in the signature.
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}
}

\examples{
showClass("IncRegFinalmodelResult")
}
\keyword{classes}
