\name{processing}
\alias{processing}
\alias{DEBUG}
\alias{errorPrintOfTheClasses}
\alias{getFormula}
\alias{getSquareBracket}
\alias{getValuesWithoutYAsDataFrame}
\alias{ownCat}
\alias{ownIfElse}
\alias{printReduction}
\alias{printScaleAndCentering}
\alias{reduceTheDataFrame}
\alias{setSquareBracket}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Miscellaneous
}
\description{
miscellaneous functions
no use for external
}


\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}

}

\keyword{ misc }

