\name{reanalyse}
\alias{reanalyse}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Reanalyse
}
\description{
starts the process afresh
}
\usage{
reanalyse(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
represents an object of the class incReg
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}

}

\keyword{ misc }