\name{scalingAndCenteringData}
\alias{scalingAndCenteringData}
\alias{scalingAndCenteringData-methods}
\alias{scalingAndCenteringData,data.frame-method}
\alias{scalingAndCenteringData,matrix-method}
\alias{scalingAndCenteringData,numeric-method}
\alias{scalingAndCenteringData,vector-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
scaling and centering
}
\description{
Scaling and centering of the dataset
}
\usage{
scalingAndCenteringData(dataSet, shouldScaling, shouldCentering, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataSet}{
input data set
}
  \item{shouldScaling}{
if the data should scaled: TRUE or FALSE
}
  \item{shouldCentering}{
if the data should centered: TRUE or FALSE
}
  \item{\dots}{
not used
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{incReg}}

}

\keyword{ manip }