\name{setAlphaValue<-}
\alias{setAlphaValue<-}
\alias{setAlphaValue<--methods}
\alias{setCollResult<-}
\alias{setCollResult<--methods}
\alias{setCsvFile<-}
\alias{setCsvFile<--methods}
\alias{setCutlineAdvZeroTest<-}
\alias{setCutlineAdvZeroTest<--methods}
\alias{setCutlineCorr<-}
\alias{setCutlineCorr<--methods}
\alias{setCutlineMulticoll<-}
\alias{setCutlineMulticoll<--methods}
\alias{setCutlineZeroTest<-}
\alias{setCutlineZeroTest<--methods}
\alias{setDataSet<-}
\alias{setDataSet<--methods}
\alias{setFinalModel<-}
\alias{setFinalModel<--methods}
\alias{setMissingValueIsCol<-}
\alias{setMissingValueIsCol<--methods}
\alias{setMissingValueTyp<-}
\alias{setMissingValueTyp<--methods}
\alias{setShouldCentering<-}
\alias{setShouldCentering<--methods}
\alias{setShouldScaling<-}
\alias{setShouldScaling<--methods}
\alias{setYname<-}
\alias{setYname<--methods}
\alias{setMultiCollResult<-}
\alias{setMultiCollResult<--methods}
\alias{setModelResult<-}
\alias{setModelResult<--methods}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Setter functions
}
\description{
Set the corresponding value for an incReg-Object
}
\usage{
setAlphaValue(object) <- value
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
represents an object of the class incReg
}
  \item{value}{
represents the corresponding value which should change or set
}
}

\author{
Alexander Entzian \email{a.entzian@alexander-entzian.de}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{getAlphaValue}}
\code{\link{incReg}}

}

\keyword{ manip }

